devdraw/drawclient: fix Not Responding issue Problem: On some computers, any program using devdraw will make the window stop responding, as soon as a key is pressed.
Devdraw.exe acts both as a GUI program, and a program
that uses the standard handles like stdin and stdout
for communication with the client. This seems to
be a bit uncommon on Windows, and on some computers
it apparently causes the problem that even if stdin
is not connected to keyboard input, but to a pipe,
it seems to interfer with the GUI's keyboard input.
When devdraw.exe waits in read(0, ...) to get 9P
messages from the draw client, and a key is pressed,
this stops the window's message loop from working
(no event is received anymore), even if fd 0 does
not refer to keyboard input.

I changed devdraw.exe, so that it now does not
use the initial pipe between the client and itself
anymore, but rather creates a new pipe and sends
a handle refering to one end of the pipe to the
client. This way I could make it run on a computer
that showed the problem previously.

Perhaps in some versions of Win32 DLLs the stdin
handle is also used by other internal mechanisms,
in case the program is using one or more windows.